Integrate plugins as sub-commands
Add FlatDirPluginSupplier, basic plugin list

Provides a temporary, naive implementation of a supplier that looks for
plugins in a specific directory, and uses said plugins to construct
toplevel subcommands in the CLI.

This is done via constructing PluginInfo by invoking the each plugin
binary's 'info' command. Failure to obtain expected output from the
command is not fatal, but will result in suboptimal description about
the installed plugin.

Also implemented a basic plugin list to make use os the same supplier.

Signed-off-by: Vui Lam <vui@vmware.com>