Add builder plugin
Fixes: #15

Enables the building of the builder plugin, to be used in turn to
compile code of other plugins.

For now, the requirement that the companion test plugin is not strictly
enforced. Set env var TZ_ENFORCE_TEST_PLUGIN=1 to turn on enforcement.
Without this flag, the building of a plugin's test plugin is skipped if
<plugin_root_dir>/test/ is absent.

The plugin is also used to publish built plugin artifacts as well as
generate go project scaffolding for the new plugin development.
TODO: update publish and scaffolding functionality.