Update builder templates with new imports
Fixes: #17

The generated project's Makefile expects the TZBIN to be set to the name
of the CLI binary located in the user's path. Its default value is
currently set to 'tz'. This convention will allow the CLI under
development to coexist with the released Tanzu CLI, typically name
'tanzu'.  We should continue to adopt said convention until the CLI
under development is released as a backward-compatible replacement of
the existing CLI.

Signed-off-by: Vui Lam <vui@vmware.com>