Remove "Describe" from the Discovery interface
A plugin is now associated with a target, so to uniquely identify a
plugin, we normally need a name-target combination.

The Describe function of the Discovery interface only took a plugin
name so if the discovery contained plugins using the same name for
two different targets, the function would not know which plugin to
return.  Since the return value is a single plugin, we didn't have
the option to return all plugins of a certain name.

We considered adding a "target" parameter to the Describe function
signature.  However, this was a little complex for a REST discovery
as its current endpoint support did not explicitly support the
specification of a target.

Since the Describe function is actually never used, we decided that
instead of spending time fixing these issues, we would simply remove it.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>