Use an env var for the plugin repo for pre-alpha
To be able to support running a new CLI but then switch to an older one,
and vice versa, we need to make sure the configuration files are
compatible between the old and new versions.

The Central Repository feature uses a different discovery source
than older CLIs so we cannot configure both at the same time and expect
to switch between CLI versions.

For the pre-alpha release we avoid the whole issue by not using the
config files for the Central Repository feature.  Instead, the plugin
discovery source for the Central Repo will be configured using an
environment variable named: TANZU_CLI_PRE_RELEASE_REPO_IMAGE

With this approach, discovery source can continue to be used by older
CLIs while the new CLI will get its source from the variable.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>