Fix version and status in plugin search/list
In the output of "plugin search", if a plugin is installed, the
installed version is now displayed instead of the recommendedVersion.

Also, in the output of "plugin search" and "plugin list", the "status"
column no longer shows "update available" but only "installed" or
"not installed" for the following reason:

Say I'm using TKG 1.6, with the Central Repo, I would see a
"update available" message because the TKG 2.1 plugins would be
available; this would be confusing.  This was not the case before the
Central Repo because only TKG 1.6 plugins would have been discovered.

The "update available" feature, will need to be evaluated to be properly
re-enabled.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>