Enable Central Repo feature by default
We want the central-repo feature to be enabled by default, but we want
a *hidden* way to disable it, just in case.  Therefore, we don't want
the feature flag to be present in the config file unless we actually
turn off the feature. This is important so that users don't think they
can turn off the feature.

However, if a feature flag is not set, it defaults to false.  This means
the central-repo feature would be disabled by default.  To work around
that, we flip the feature flag to be a way to *disable* the central
repo feature instead of enabling it like before.

So the feature flag becomes "no-central-repo-test-only":
- false -> central-repo feature enabled
- true  -> central-repo feature disabled

The suffix "test-only" is meant to make it even clearer that this
feature should not be turned off.

Unit tests affected by this change (now running with the central repo
feature enabled) have been adapted.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>