Add filtering to the SQLite-backed OCI discovery
The SQLite-backed OCI discovery now has "criteria" field which allows
to only discover plugins that match that criteria.

At the same time the PluginInventory now has a "PluginInventoryFilter"
field which allows to only List() plugins respecting the filter.
When using the PluginInventory, the OCI discovery converts its criteria
into a PluginInventoryFilter to only request the proper plugins.

Using this filtering system, it is now possible to limit the amount of
fetched plugin data to a more targeted subset.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>