Add support for sandbox testing
The environment variable ADDITIONAL_PLUGIN_DISCOVERY_IMAGES_TEST_ONLY
can be used as a comma separated list of OCI images in the same format
as the Central Repo.

Plugins found in those repos can be installed by the CLI.  This allows
to setup an sanbox repos, set the variable and then install from either
the sandbox repo or from the Central Repo.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>