Keep track if a plugin is installed from a context (#97)
Fix #96

With the Central Repo feature, this allows "plugin sync" to properly
store in the plugin catalog, which plugins are context-scope and which
are standalone.

Before the Central Repo feature, plugin installation was done by
discovering every plugin which allowed to know which plugins were
context-scope.  With the Central Repo, we install from one global
central repository which doesn't (and shouldn't) provide the scope
of the plugin. The criteria to decide the scope of a plugin is if it was
installed manually (`tz plugin install` or through a context
`tz plugin sync` or `tz context create/use`).

Signed-off-by: Marc Khouzam <kmarc@vmware.com>