Teach pluginsupplier to prioritize context plugins
It is possible to install a plugin both as standalone and context-scope
at the same time.  This can be done by simply creating a context that
installs a plugin and then installing the same plugin by hand using
"plugin install", or vice versa.

When such a situation occurs, the catalog has knowledge of both plugin
instances (which may even be at different versions), but the CLI can
only use one of them.  The context-scope plugin is given priority as it
is mapped to a context and therefore more important.

If a user wants to force a different version of a context plugin they
can first uninstall the plugin then install the other version by hand.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>