Align yaml and json parsers for the catalog
Fixes #118

When the CLI was transferred over we moved from using a json parser to
a yaml parser.  for the new CLI to be able to read old CLI's catalog
it is important that the fields be named the same, including case
sensitivity.

This commit adds the proper annotations for 'yaml' to the different
go structures that are used by the catalog, so that the parsing matches
the original format.

A unit test is also added to keep testing compatibility as the CLI
continues to evolve.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>