Support installing a plugin as both standalone and context at the same time (#117)
* Teach pluginsupplier to prioritize context plugins

It is possible to install a plugin both as standalone and context-scope
at the same time.  This can be done by simply creating a context that
installs a plugin and then installing the same plugin by hand using
"plugin install", or vice versa.

When such a situation occurs, the catalog has knowledge of both plugin
instances (which may even be at different versions), but the CLI can
only use one of them.  The context-scope plugin is given priority as it
is mapped to a context and therefore more important.

If a user wants to force a different version of a context plugin they
can first uninstall the plugin then install the other version by hand.

* Fix "plugin delete" for two instances of a plugin

It is possible that the catalog contains two entries for a plugin:
a context-scope installation and a standalone installation.  We need to
delete both in this case.

If all matched plugins are from the same target, this is when we can
still delete them all.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>