Fix `target` inconsistency between InsertPlugins and GetPlugins plugininventory APIs
Store full-form of targets ('kubernetes' & 'mission-control') in the database

Currently, the InsertPlugin stores full-form of the target values such
as `kubernetes` and `mission-control` to the database. However, the
plugin get filtering mechanism expects short-form of the target (`k8s`
and `tmc`) to be stored in the database.

Storing shorter format requires additional conversion to be made while
storing and reading the target information hence updating the logic to
make it more generic.

Verify the fix using unit tests