Add prefix TANZU_CLI_ to newly added env vars
This is meant to avoid any collisions with other uses of a variable with
the same name (no matter how unlikely).
The two modified variables have been moved to `env_variables.go` where
they should have been created from the start.

As for the constant meant to bypass the pre-release repo image, as it is
not a variable name but a value, it does not need the TANZI_CLI_ prefix
and it does not belong in the same file.  It was therefore moved to the
pluginmanager; note this constant will be removed in the near future
anyway.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>