Hide the "plugin source" commands
For the alpha.0 release of the CLI, plugin discovery sources are not
used to configure the Central Repository.

This commit hides the commands to avoid any confusion.
These commands were not really used before anyway, so this should not
even be noticed by alpha.0 users.

Plugin discovery sources will be enhanced and re-enabled after the
alpha.0 release.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>