Update OS Package recipes for the new repo
Some updates were missed when we moved the OS Packages recipes to this
CLI repo.

Also, handle '-' character for RPM packages in preparation for an alpha
pre-release.

Also add an extra directory "tanzu-cli" under the rpm directory; this
will allow to include other rpm packages in the same directory
structure.

We also prefix the code name "jessie" to be "tanzu-cli-jessie".
This will allow to include other apt packages under the same directory
structure but using a different name.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>