Use caching for plugin discovery images
Fix #147

Because of discovery image signature verification which can take a few
seconds, the performance of commands such as "plugin search" has become
too slow.  To improve this, we start using caching of the DB images.

The solution chosen has two aspects:
1. Check the cache validity and re-use the image when appropriate.  When
   re-using the image, there is no need to verify its signature which
   speeds up the operation significantly.
2. When the cache needs to be updated, we notify the user that the DB is
   being updated which can take a few seconds. This makes the UX much
   smoother in itself.

With this commit, the cache directory structure stays the same except
that a new file is added in the same location as where the DB image
is downloaded.  This file is named "digest.<imageHash>" and is empty.
It serves the purpose of indicating the hash value of the cached DB
image.  For example, we could have:

~/.cache/_tanzu/plugin_inventory
├── default
│   ├── digest.b160d48e089d01b60303e4ff91f8858bd0c81db3d56f3ba0a21aa29091c18c79
│   └── plugin_inventory.db
└── test_0
    ├── digest.89edcfa8a1bf54744dfef48da1b77c82697f278a4170eec9693b0af7f3c7d049
    └── plugin_inventory.db

The first directory ("default") represents the default plugin discovery
image, while in this example, the second entry represents an additional
test image added through the use of the environment variable
TANZU_CLI_ADDITIONAL_PLUGIN_DISCOVERY_IMAGES_TEST_ONLY.

To make sure we don't use a stale cache, whenever we need to use a DB
image, we always first download its most recent digest hash and compare
its value with the "digest.<imageHash>" file name.  If they match, it
means we can re-use the cached DB.  If they don't match, we need to
refresh the cache.  In the latter case, we must also remove the old
digest file.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>