Update main.go template to avoid import-not-used
A recent change to introduce a ""FIXME<<<<" string instead of using
types.TargetUnknown is meant to better explain to developer that the
target has to be explicity set, but after this change, the error
types imported and not used becomes more confusing and does not point
to where the fix is needed.

Until we move away from needed developer to fix the generated code
update the template to eliminate the error.

before:
cmd/plugin/myplugin/main.go:6:2: imported and not used:
"github.com/vmware-tanzu/tanzu-plugin-runtime/config/types"
make: *** [plugin-build-darwin-amd64] Error 1

slightly better:
2023-04-06T09:09:31-07:00 [x] building plugin "cmd/plugin/myplugin" failed - exit status 1:
[x] : invalid PluginDescriptor specified: plugin "myplugin": target is not valid

Signed-off-by: Vui Lam <vui@vmware.com>