Delete entries using unknown target
Plugins aimed at a "global" target (not "k8s" nor "tmc") use a target
named "global", but used to use the unknown/unspecified target which was
just the empty string.

If an plugin is installed with the "global" target, it now replaces any
installation that uses the unknown target and vice versa.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>