Don't get hidden (deactivated) plugins from the DB
This commit adds a "IncludeHidden" boolean field to the plugin filter
struct and updates the SQLite query to respect that field.

Furthermore, the pluginfilter argument to GetPlugins() when nil, is
replaced with an empty filter. This allows the empty value of such a
filter to ignore hidden plugins automatically since the field
IncludeHidden will be false by default.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>