Provide a means to list all versions of a plugin (#211)
Fixes #134

This commit introduces a "--name" flag to "plugin search".  Using this
flag it is possible to limit the search to plugins with a specific name.

This commit also adds a "--show-details" flag to "plugin search"
(instead of the orignally intended "--list-versions").  Using the
"--show-details" flag, it is possible to get more information about each
plugin, including all versions that are available.

This commit also adds a "--target" flag to "plugin search".  Using this
flag, it is possible to limit the search to plugins of the specified
target.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>