Use log.Warningf() insead of fmt.Printf() (#226)
`log.Warningf` prints to stderr while `fmt.Printf` stdout. For the case
in question, we are processing an error but the output will be on stdout
for one part and stderr for the other.

Also, `log.Warningf` automatically adds a newline at the end while
`fmt.Printf` does not.  This is an issue in this case as we will have
multiple lines of output.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>