Replace direct access with library config funcs (#225)
The configuration support provided by tanzu-plugin-runtime has functions
that allow to get and set discovery sources.  The CLI should not access
the raw config data but instead use these functions.

This commit replaces direct access with these function calls.
This modification causes no change in behaviour.

Note that unit tests are actually removed with this commit.
Those tests were verifying code that is now provided by the
tanzu-plugin-runtime library.  The test functions were not removed
as they should be modified to test the actual cobra command logic.
This will be done in a follow-up change.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>