Use discovery sources for plugin discoveries
This commit removes the use of TANZU_CLI_PRE_RELEASE_REPO_IMAGE which
was a temporary approach.

Instead, the central discovery is automatically created in the config
file as a "discoverySource" and "tanzu plugin source update" can be used
to override this default value.

This commit also updates the "tanzu plugin source" family of commands.
With the central discovery feature enabled:
1. "tanzu plugin source add" is removed
2. "tanzu plugin source update" no longer has a "--type/-t" flag
3. "tanzu plugin source init" is added

Signed-off-by: Marc Khouzam <kmarc@vmware.com>