Provide configuration variable to access deactivated plugins/groups (#219)
* Adds env var to access deactivated plugins/groups

To test their plugins before activating them in the central repo,
publishers can now set the environment variable
TANZU_CLI_INCLUDE_DEACTIVATED_PLUGINS_TEST_ONLY=true

This tells the CLI to include any deactivated plugins and groups as if
they were active.

When creating a plugin group, if hidden plugins are to be allowed in
the new group, the TANZU_CLI_INCLUDE_DEACTIVATED_PLUGINS_TEST_ONLY must
also be set.

* Updates tests to handle hidden plugins/groups

It was becoming increasingly difficult to test
`listPluginsFromInventory()` and `listGroupsFromInventory()` now that
we had to consider hidden plugins.  The stubbed inventory logic would
have to be pretty advanced to be able to properly filter out hidden
plugins when needed.

This commit takes a different approach to testing the `oci_dbbacked.go`
file.  We use the fact that we don't need to test the logic implemented
by the inventory stub but only the fact that the stub should be called
with the correct plugin filter. Instead of returning a filtered list
of plugins, the stub returns the plugin/group filter by using the error
object; the test can then verify that the filter is appropriate to
provide the type of plugins/groups that is expected.

This type of testing is sufficient since there are other tests for the
real plugin inventory code which is its own package.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>