Remove unit test that was no longer relevant
With PR #232 the trusted registries include the configured plugin source.
The default central repo URI is no longer automatically trusted but only
if it is configure as a plugin source (which happens by default).

Therefore, the unit test is superceded by the unit tests of the same
file that verifies the trusted registries for configured plugin source.

Furthermore, the test that this commit removes would fail locally when
the plugin source configured is not the default central repo.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>