Add --name to plugin group search (#267)
* Add --name to plugin group search

This commit allows filtering by name for the "tanzu plugin search"
command. To achieve this the commit improves the discovery mechanism
to have an explicit GroupDiscovery which can be created with a
GroupCriteria. This approach is the same as what we do for plugin
discovery.

* Properly handle invalid group identifier

Now that we use a group criteria to request a list of groups, it is
important to notice if the groupId specified by the user is invalid
before setting the filter to empty values, which would mistakenly
return all groups.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>