Verify plugin source before accepting to update it
This commit also removes the log.Fatal() when fetching the plugin
inventory OCI image.  It is not possible to use unit tests on a function
that calls log.Fatal(), and since this commit now verifies the validity
of the OCI image in the createDiscoverySource() function, I wanted to
be able to have some unit tests make sure the verification was correct.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>