Boolean flags must use the `=` format
Also, there cannot be any extra spaces between arguments for a command
because we split the command string by spaces; if there are extra spaces
that creates extra empty arguments.

To help us detect any such future error, this commit adds a printout of
the list of arguments used when executing an E2E tanzu command.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>