APT packages now use a binary built locally
This means that when such packages are built locally on a developer's
machine, they will not be SRP-compliant since they no longer use the
SRP-compliant binary from Github.

Building these packages should be done using an SRP-compliant
environment and are meant to be done at the same time as when the
SRP-compliant binaries are built.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>