Allow plugin inventory cache to work on Windows
Fixes #392

On Windows, trailing `.` are ignored in file names. But we used a file
`metadata.digest.` with a trailing `.` when not using an airgapped image
which is the most common case actually.

Since the trailing `.` gets ignored on Windows, the CLI never finds the
`metadata.digest.` file and always thinks the cache is invalid.

This commit replaces the use of `metadata.digest.` with
`metadata.digest.none`.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>