A global plugin is installed at the root over k8s
If a global plugin is installed and a k8s plugin with the same name is
also installed, the global plugin should be the one that is installed
at the root level.

This change allows a plugin dev to move a plugin from the k8s target to
the global target.

For example, the "telemetry" plugin is a k8s plugin at the moment.
If it is later published as a "global" plugin and that the user installs
both versions (or installs the new one with the old one already present)
this commit makes sure that when calling `tanzu telemetry` it will be
the new "global" plugin that will be called.  If the other version is
still installed, it will remain accessible under `tanzu k8s telemetry`
until the user uninstalls it.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>