Build the RPM repository incrementally
A requirement that we have is that our repo is on a GCP bucket
(at least for the moment).  Therefore, all existing versions of the CLI
packages are on the bucket and not on the machine where we build the
repo.  We don't want to have to download every package from the bucket
to be able to build a new repo as this would take more and more time
as we accumulate versions.  Instead we want to build the repo
incrementally by simply adding a new package version to it.

After investigation, the best we I found to do this was to download the
existing repo metadata, locally create empty packages for the existing
packages, and the rebuild the metadata with using:
  createrepo --update --skip-stat

This approach re-uses the previous metadata but adds to it the new
package versions.

This commit also numbers pre-releases appropriately for RPM.
In case we ever put pre-release packages in a repo along with final
releases, it is important that the pre-releases be numbered as such
for RPM to prioritize the final releases.

The commit also does a little cleanup to use a loop per architecture.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>