Build plugins for ARM64
We start building the 'builder' and 'test' plugins for ARM64.
We also default new plugin projects to also building for ARM64.

The builder plugin will publish the ARM64 binaries and include them in
the plugin inventory DB if they are present.  If they are not present,
the builder plugin will ignore them with a printout.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>