ActiveHelp: cmd short desc when completing cmds/args
Setup a generic ValidArgsFunction to add a command's 'short' string
as activeHelp when doing arg completion for any command.

Doing something like "tanzu context <TAB>" will show normal shell
completions but will now also print the short help for "tanzu context".

Note that activeHelp only gets shown for `bash` and `zsh`.

```
$ tanzu context <TAB>
Command help: Configure and manage contexts for the Tanzu CLI
--
create  -- Create a Tanzu CLI context
delete  -- Delete a context from the config
get     -- Display a context from the config
list    -- List contexts
unset   -- Unset the active context so that it is not used by default.
use     -- Set the context to be used by default
```

The prefix added before the help text is to deal with situation like
the following:

$ tz context create tmc --endpoint unstable.tmc-dev.cloud.vmware.com <TAB>
Command help: Create a Tanzu CLI context

Without the prefix the text "Create a Tanzu CLI context" could be
interpreted as an action the user must do, instead of the description of
the command.

Because this change would affect most of the shell completion unit tests,
each batch of such test was updated to disable this feature using
TANZU_ACTIVE_HELP=no_short_help.

A unit test specific to this feature was added in root_test.go.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>