Cleanup discovery source commands
This commit does the following:
1. remove the code that ran if the Central Repo feature was disabled, as
   we don't support turning off the Central Repo anymore
2. move the creation of flags closer to the commands themselves
3. mark the --uri flag as required for "plugin source update"; this was
   forgotten when we implemented the central repo feature; also update
   the usage text to include --uri since it is required

Increase coverage for "plugin source" tests

This commit adds tests for "plugin source update" and
"plugin source delete". The coverage for the discovery_source.go file
goes up from 55.5% to 93.7%

It also fixes the test for "plugin source list" which was not using a
temporary configuration; not using a temporary configuration makes it
that the tests will change the real CLI configuration for the user that
runs the tests.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>