Cleanup the use of feature flags
This commit removes the code that was no longer run due to feature flags
not being supported anymore.  The unit tests are also updated to match
this reality.

Furthermore, now that we don't support turning off the Central repo of
plugins feature, the only OCI discovery that is supported is the
"DBBackedOCIDiscovery", so this commit removes the "OCIDiscovery" type
and its supporting code.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>