Fallback to installing AMD64 plugins on darwin (#491)
Although the CLI is available for Darwin ARM64, the vast majority of
plugins are not, and therefore an ARM64 CLI will not be able to install
most plugins.

Thanks to Apple's Rosetta emulator available on Darwin ARM64 machines,
it is possible to run AMD64 binaries.  This is currently how the CLI
and its plugins are being run.

This commit teaches a Darwin ARM64 CLI that if a plugin is not available
for Darwin ARM64, it should instead install the Darwin AMD64 version.
With this approach, it now become possible to use a Darwin ARM64 CLI to
its full potential.

Note that this approach cannot be used for Linux as there is no standard
emulator for AMD64 binaries.  For Linux, plugins will need to be
published for ARM64.

Unit tests have been added for this new feature.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>