Complete valid values for --target
When using the --target flag, the shell completion logic is now smarter
and only specifies the valid targets for the current command-line.
For example:

$ tanzu plugin install cluster --target <TAB>
k8s  -- For interactions with a Kubernetes cluster
tmc  -- For interactions with a Tanzu Mission Control endpoint

Notice the 'global' target was not suggested because there is no
'cluster' plugin for it.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>