Implement reprompt of updated EULA
This change adds the capability to reprompt the user to accept an
updated EULA, should it become available. At some point, the version of
the EULA associated with the CLI will be versioned using semver. Once
that happens, a reprompt will occur even if it was previously registered
that a EULA was accepted if a EULA version compatible (matching
major.minor) with the CLI's EULA version is not found is the list of
accepted versions in the config file.

Note that as of this change, the CLI has not designated a specific
version, hence this compatibility check is not activated.

Signed-off-by: Vui Lam <vui@vmware.com>

(cherry picked from commit 59195116c7d62ccd508a4ce327682d24df283870)