Complete valid values for --target (#555)
When using the --target flag, the shell completion logic is now smarter
and only specifies the valid targets for the current command-line.
For example:

$ tanzu plugin install cluster --target <TAB>
k8s  -- For interactions with a Kubernetes cluster
tmc  -- For interactions with a Tanzu Mission Control endpoint

Notice the 'global' target was not suggested because there is no
'cluster' plugin for it.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>
(cherry picked from commit ad0fd0119b452ea7e317ca2161a836e0aee18218)