Include aliases in the main help
Fixes #564

The CLI has two types of help format:
1- for the main command
2- for sub-commands

Type 1 is for the help of the "tanzu" command itself, but also for
the help of the targets: "tanzu kubernetes" and "tanzu tmc".  We do this
because such targets are commands that have plugins as sub-commands, and
plugins can choose in which help group they want to be shown.

Type 2 is for all core sub-commands of the CLI.

This commit adds an "Aliases:" section to help type 1 just like there is
already for type 2.  If there are no aliases, that section is not shown,
therefore, the help of the root command will not be changed since the
root command does not have aliases.  On the other hand, with this commit
the two "tanzu kubernetes" and "tanzu tmc" help texts will show the
"k8s" and "tmc" aliases, respectively.

This should help the user realize such aliases are available.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>