Don't disable flag parsing on the root command (#600)
By disabling flag parsing on the root command, we tell Cobra not to
parse flags when running the root command.  Since there are are
currently no flags on the root command this is not a problem.
Furthermore, even with say, a global flag on the root command, this
situation would still not be a problem because the root command is not
actually Runnable itself, so flags would not be executed on the root
command itself, but instead the DisableFlagParsing field would be read
from the executable sub-command used.

However, there is an issue with not parsing flags on the root which is
that it prevents Cobra's shell completion logic to complete the
--help/-h flags on the root command.  This is because all flag name
completion becomes the responsibility of the program itself, and the
program did not itself define --help/-h.  This happens starting with
Cobra v1.8.0.

We can see this by running `tanzu __complete -` and noticing the
--help/-h flags are not suggested when using cobra 1.8.0.

The reason flag parsing was disabled on the root seems to be to allow
plugins to handle their own flags.  However, this is already done by each
individual plugin sub-command created by `GetCmdForPlugin()` which does
disable flag parsing already.  There is therefore no reason to disable
flag parsing on the root command.

Signed-off-by: Marc Khouzam <kmarc@vmware.com>