Refresh the plugin DB on 'plugin source init'
Not refreshing the plugin inventory DB on "plugin source init" can
cause confusion for some situations.  For example, shell completion,
which only uses the inventory cache, would continue using the old cache
after a "plugin source init"; similarly, the installation of the
essential plugins would not get updated until the DB was refreshed
through another command.

Furthermore, this becomes even more important as we plan to improve
the DB cache to not check the digest for a certain period of time; in
such a case, after running a `plugin source init`, the DB could be stale
for all commands.

Signed-off-by: Marc Khouzam <marc.khouzam@broadcom.com>