Use the local "builder" version
When running the tests for the sample-plugin, the "builder" plugin needs
to be installed. However, it may happen that there was a change in the
sample-plugin Makefile and the builder plugin code.  This means we must
use the local "builder" version and not the one currently published to
the central repository.

This commit does this by:
1. building the "builder" plugin under bin/
2. using bin/builder to build the "builder" plugin so it can be
   installed under the tanzu CLI

Signed-off-by: Marc Khouzam <marc.khouzam@broadcom.com>