Hide a target command that has no plugins
When a target (`k8s` or `tmc`) has no plugins, this commit hides the
command for that target.  The target may never be applicable to some
users, so we try to avoid complicating things by hiding it.

Note that we don't completely remove the command for such a target
because it is possible that savvy users expect it to be there even
before they've installed plugins.

Further, when no plugins are installed for a target:
- `tanzu <target>` will now print the help including a note
- `tanzu <target> invalid` will now print the help including a note
- `tanzu <target> -h` will now print the help

Previously, the above cases would only print, e.g., for TMC:
"Tanzu CLI plugins that target a Tanzu Mission Control endpoint"

The commit also removes the "Available command groups:" section of the
help of a target if no plugins are available for that target.

Signed-off-by: Marc Khouzam <marc.khouzam@broadcom.com>