Document running "brew tap vmware-tanzu/tanzu" (#607)
After investigation and different tests writing a formula for
homebrew-core it is clear that we cannot provide our SRP builds of the
CLI through the homebrew-core community tap; instead, the build of the
CLI must be done by homebrew's CI. Beyond the benefits of using an SRP
build, using the exact SRP build of the CLI is important for certain
authentication workflows. Therefore, we cannot use the homebrew
community tap.

The main issue we were trying to solve with the use of the homebrew
community tap is that it is cumbersome to use
`brew install vmware-tanzu/tanzu/tanzu-cli` and it would be much nicer
to perform brew operations on `tanzu-cli` instead of
`vmware-tanzu/tanzu/tanzu-cli`.

As an alternative we simply update our documentation to run the
`brew tap vmware-tanzu/tanzu` command explicitly which then allows to
run `brew install tanzu-cli` directly as well as other operations
such as `brew upgrade tanzu-cli`.

Note that this approach does not help with "discovery". A new user
cannot simply type `brew install tanzu-cli` but first needs to find
which tap must be used. However, this is a small hurdle as a google
search can quickly point to vmware documentation on how to install the
CLI.

Signed-off-by: Marc Khouzam <marc.khouzam@broadcom.com>